refactor(ChatSupport): 1831行モノリスを hooks + components に分割
ChatSupport.tsx (1831行) を以下の構成に分割:

hooks/
  useChatState.ts    (108行) — isOpen/messages/scrollRef/clearChat
  useChatConfig.ts   (136行) — config/settings UI状態/API設定
  useChatMessage.ts  (326行) — 送信/FAQ/AI/セマンティック検索
  useChatShortcuts.ts(109行) — グローバルキーボードショートカット

components/
  ChatHeader.tsx       (135行) — タイトル/ツールバー/モデル名
  ChatInput.tsx         (81行) — 入力欄/送信ボタン
  ChatMessageList.tsx  (255行) — メッセージ/サジェスト/タイピング
  ChatPageContextChip.tsx(69行) — ページ解説トリガー
  ChatSettings.tsx     (381行) — API設定/モデル/ショートカット

ChatSupport.tsx: 1831行 → 273行（コンテナのみ）

sdpf-theme(222行) / Matlens(177行) との足並み揃えの一環。
